Demonstrating resource filtering using Maven
Setting a property in application.properties file with @value@ will allow the property to be injected by Maven during the build process if the intended value if specified in the pom file. In this example, a placeholder named 'jwt-uri' is replaced during the Maven build process with the string specified in the pom file under the <jwt-uri></jwt-uri> tag.